Bug #30705 - Add conditional file filters.
I effectively merged these new filters into the existing And and Or
filters, rather than having multiple flavours of them. In practice, this
means that I added the contructors from the original filters to these new
ones, and then replaced the old with the new. All of the pre-existing
tests work with the resulting filters, along with all the new tests
supplied with the enhancement request. Thanks to Steven Caswell for the
code.

I also had to exclude the abstract test cases in the Maven build file, so
that JUnit wouldn't try to instantiate and run them, and regenerate the
Ant build file.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/io/trunk@140618 13f79535-47bb-0310-9956-ffa450edef68