Add an IfStatement to check the return value of lockFile.delete().
This statement returns a value that is not checked.
The return value should be checked since it can indicate an unusual or unexpected function execution.
The statement returns false if the destination file could not be successfully deleted (rather than throwing an Exception).
If you don't check the result, you won't notice if the statement signals an unexpected behavior by returning an atypical return value.
http://findbugs.sourceforge.net/bugDescriptions.html#RV_RETURN_VALUE_IGNORED_BAD_PRACTICE