Remove outdate javadoc references to TestCase#setUp()
In JUnit 3, in order to run some code before each test, you would
extend junit.framework.TestCase and override its setUp() method.

In JUnit 4 this is no longer the case, and you simply need to annotate
the method with org.junit.Before, as is done in the classes in
question.

This patch removes old javadoc comments that reference the TestCase
method which are no longer relevant, and are probably outdated remains
from the days the project used JUnit 3.