Standardize org.junit.Assert imports
The de-facto standard for using org.junit.Assert methods in the
project is to statically import its method.
However, some places use the class directly and call methods on it.
This patch standardizes the approach, and makes sure all the usages of
org.junit.Assert statically import methods in order to have a
consistent coding standard throughout the project's tests.