Backport `Objects.checkFromIndexSize` to `IOUtils` (#790)
* Backport `Objects.checkFromIndexSize` to `IOUtils`

Java 9 introduced `Objects.checkFromIndexSize` for validating the offset/length ranges used in `read` and `write` operations.

This PR adds an equivalent method, `IOUtils.checkFromIndexSize`, to provide the same validation in Commons IO. It simplifies range checks in stream implementations and makes the utility available to projects that still target Java 8.

* Add changelog entry

* fix: apply `IOUtils.checkIndexFromSize` to all streams

* fix: failing tests

* fix: inline static import

* fix: add usage examples to Javadoc

* fix: Javadoc for JDK 8 and 11

* fix: remove `@Override` from code snippet

There is no way to insert `@Override` in a Javadoc snippet without triggering a warning in either Java 8 or 21.

* fix: `ClosedReader` Javadoc and tests

* fix: revert `BrokenReader` and add tests

The tests are improved to ensure that the given exception is thrown **before** parameter validation.

* fix: remove `BrokenReader.read(char[]/CharBuffer)` overrides

* fix: revert `BrokenWriter`

Revert `BrokenWriter` and add test cases to ensure the exception is thrown **before** parameter validation.

* fix: Checkstyle failures

* fix: revert `ClosedWriter` and add tests

* fix: revert `FilterCollectionWriter` changes

* fix: fix `NullAppendable` Javadoc

* fix: fix `NullOutputStream` Javadoc and enhance tests

* fix: fix `NullWriter` Javadoc and enhance tests

* fix: `NullInputStream` Javadoc

* fix: `ClosedOutputStream` Javadoc

* fix: `StringBuilderWriter` behavior and Javadoc

* fix: restore `NullWriter` comment

* fix: `checkFromToIndex` logical or

* fix: document `IndexOutOfBoundsException`

* fix: `ClosedInputStream` and `ThresholdingOutputStream` javadoc